+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Disallow duplicate player party titles
 - Fix duplicate interact entries on remote sometimes
 - Fix physics server not focusing gui's when hovering with mouse after using keyboard to navigate menus
+- Fix websocket transport not working on windows
 
 ## [0.8.4] 2026-08-19